Fixed post proposal banner width
